you-predict-core/pyproject.toml
[project]
name = "you-predict-core"
version = "0.1.0"
description = "YouTube video performance prediction platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.8.0",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-tasks>=2.17.0",
    "google-api-python-client>=2.166.0",
    "youtube-transcript-api>=1.0.0",
    "scikit-learn>=1.6.0",
    "xgboost>=2.1.0",
    "pandas>=2.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.0",
]

[tool.ruff]
target-version = "py313"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/models/raw.py" = ["N815"]  # camelCase fields must match YouTube API

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = [
    "google.*",
    "googleapiclient.*",
    "youtube_transcript_api.*",
    "xgboost.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
